Blockquotes-in-list and running blockquotes fixes
The original Markdown implementation supports "running blockquotes": if
any line in a paragraph start with the '>' character, that line and all
the subsequent ones are split from the paragraph and become a
blockquote. This is inconsistent with the list behavior (list don't
start mid-paragraph).

Additionally, if a blockquote happens within a non-block list item (e.g.
a standalone item or an item in a sequence of items not separated by
empty lines), mismatched markup is generated, with interleaved
'blockquote' and 'li' tag pairs because Markdown starts thinking it's in
span mode, and then reparses the span-mode output in block mode.

Blockquote-in-list detection is solved by letting the list item
processor check for existence of >-starting lines in the whole item.
Since this is inefficient, we allow the user to disable running
blockquotes, in which case blockquotes cannot start mid-paragraph
(consistently with the list behavior) and the blockquote-in-list
detection is much more efficient.