Changed colons to underscores and added anchors w/ names for footnotes
Posterous (and possibly other CMS systems) does not like colons in the "id" or "name" attributes of html elements. Also, it strips out "id" attributes for most elements (including the "li" element of the MMdown footnote). So this commit adds "a name=fn_footnoteid" elements to the footnote "li" elements, as well as changing the footnote name/id to "fn_footnoteid" rather than "fn:footnoteid".